Allow using a function or async function for `proxyRes` and `end`
Previously we were only emitting events, which meant
you couldn't do anything in the event to modify the response.
It also meany you couldn't call async functions in the event to,
say, conditionally return responses.

If the Promise returns null, the request and repsonse will be aborted.
Note that for `end` returning null does nothing.